chore: renamed MacActivityIcon file
