+++ apps/meteor/ee/client/omnichannel/components/MacActivityIcon/MacActivityIcon.tsx
